Fix device management bug
AutoModel wasn't loading because moving data out of meta tensors (which from_pretrained creates) is not possible. The reason is that StripedHyena's class definition has intricate device management logic. Ripping out that logic from model.py turned out to be surprisingly painless: just remove a for loop that distributes layers across GPUs, and two instances of with(device):

Now we have a layer naming bug, which is fantastic progress